remove the DsWriteInfo class
