+++ spinn_front_end_common/interface/ds/__init__.py
@@ -15,9 +15,7 @@
 
 from .data_specification_targets import DataSpecificationTargets
 from .data_row_writer import DataRowWriter
-from .ds_write_info import DsWriteInfo
 
 __all__ = [
     "DataRowWriter",
-    "DataSpecificationTargets",
-    "DsWriteInfo"]
+    "DataSpecificationTargets"]
