+++ spinn_front_end_common/interface/ds/data_specification_targets.py
@@ -141,15 +141,28 @@ class DataSpecificationTargets(MutableMapping):
                 cores_to_mark.append((x, y, p))
         self._db.ds_mark_as_system(cores_to_mark)
 
-    def set_info(self, x, y, p, info):
+    def write_set_info(self, x, y, p, info):
         """ Sets the info for the core x, y, p
 
         :param int x: core x
         :param int y: core y
         :param int p: core p
-        ::param DataWritten info:
+        :param DataWritten info:
         """
         self._db.set_write_info(x, y, p, info)
+
+    def get_write_info(self, x, y, p):
+        """
+        gets the info for the core x, y, p
+
+        :param int x: core x
+        :param int y: core y
+        :param int p: core p
+        :return: dict with the keys
+            ``start_address``, ``memory_used`` and ``memory_written``
+        :rtype: dict(str,int)
+        """
+        return self._db.get_write_info(x, y, p)
 
     def set_size_info(self, x, y, p, memory_used):
         """ Sets the size info for the core x, y p.
