+++ unittests/interface/ds/test_ds_write_info.py
@@ -17,7 +17,6 @@ import unittest
 from spinn_machine.virtual_machine import virtual_machine
 from spinn_front_end_common.interface.config_setup import unittest_setup
 from spinn_front_end_common.utilities.utility_objs import DataWritten
-from spinn_front_end_common.interface.ds.ds_write_info import DsWriteInfo
 from spinn_front_end_common.interface.ds import DataSpecificationTargets
 
 
@@ -30,25 +29,22 @@ class TestDsWriteInfo(unittest.TestCase):
         check = dict()
         machine = virtual_machine(2, 2)
         dst = DataSpecificationTargets(machine)
-        asDict = DsWriteInfo(dst.get_database())
         c1 = (0, 0, 0)
         foo = DataWritten(123, 12, 23)
-        asDict.set_info(*c1, info=foo)
+        dst.write_set_info(*c1, info=foo)
         check[c1] = foo
-        self.assertEqual(foo, asDict.get_info(*c1))
+        self.assertEqual(foo, dst.get_write_info(*c1))
 
         c2 = (1, 1, 3)
         bar = DataWritten(456, 45, 56)
-        asDict.set_info(*c2, info=bar)
+        dst.write_set_info(*c2, info=bar)
         check[c2] = bar
-        self.assertEqual(bar, asDict.get_info(*c2))
+        self.assertEqual(bar, dst.get_write_info(*c2))
 
-        self.assertEqual(2, len(asDict))
+        for key in check:
+            self.assertEqual(check[key], dst.get_write_info(*key))
 
-        for key in asDict:
-            self.assertEqual(check[key], asDict.get_info(*key))
-
-        for key, value in asDict.items():
+        for key, value in dst.items():
             self.assertEqual(check[key], value)
